requirements.txt
gTTS==2.2.4
playsound==1.3.0
python-dotenv==0.20.0